Refactor so we don't have one big conking component
