+++ plugins/visual-qontract/src/components/ChangelogFetchComponent/ChangeTypes.tsx
@@ -1,0 +1,11 @@
+export type Change = {
+  commit: string;
+  merged_at: string;
+  change_types: string[];
+  error: boolean;
+  apps: string[];
+};
+
+export type ChangeTableProps = {
+  changes: Change[];
+};
