+++ plugins/visual-qontract/src/components/ChangelogFetchComponent/ChangelogFetchComponent.test.tsx
@@ -1,9 +1,8 @@
 import React from 'react';
 import { render, screen, fireEvent, waitFor } from '@testing-library/react';
 import '@testing-library/jest-dom';
-import { DenseTable } from './ChangelogFetchComponent';
+import { ChangeTable } from './ChangeTable';
 import { MemoryRouter } from 'react-router-dom';
-import * as ReactRouterDom from 'react-router-dom';
 
 // Sample data for testing
 const testData = [
@@ -38,7 +37,7 @@ describe('DenseTable component', () => {
   it('renders the table with correct columns and rows', () => {
     render(
       <MemoryRouter>
-        <DenseTable changes={testData} />
+        <ChangeTable changes={testData} />
       </MemoryRouter>,
     );
 
@@ -63,38 +62,41 @@ describe('DenseTable component', () => {
     expect(screen.getByText(/Oct 26,? 2024/)).toBeInTheDocument();
   });
 
-
   it('clears the search text when the clear button is clicked', () => {
     render(
       <MemoryRouter>
-        <DenseTable changes={testData} />
+        <ChangeTable changes={testData} />
       </MemoryRouter>,
     );
-  
+
     const searchInput = screen.getByPlaceholderText('Search');
     fireEvent.change(searchInput, { target: { value: 'Update' } });
     expect(searchInput).toHaveValue('Update');
-  
+
     // Locate the clear button within the search input field
     const clearButton = screen.getByRole('button', { name: '' }); // Using an empty name as placeholder might not have an aria-label
     fireEvent.click(clearButton);
     expect(searchInput).toHaveValue(''); // Ensure search text is cleared
   });
-  
+
   it('removes a filter when the "x" icon is clicked on a filter pill', async () => {
     render(
       <MemoryRouter initialEntries={['/?filters=type%3AUpdate,app%3AApp1']}>
-        <DenseTable changes={testData} />
+        <ChangeTable changes={testData} />
       </MemoryRouter>,
     );
-  
+
     // Ensure both filter pills are present initially in the filter box
-    expect(screen.getByTestId('active-filter-app-pill-App1')).toBeInTheDocument();
-    expect(screen.getByTestId('active-filter-type-pill-Update')).toBeInTheDocument();
-  
+    expect(
+      screen.getByTestId('active-filter-app-pill-App1'),
+    ).toBeInTheDocument();
+    expect(
+      screen.getByTestId('active-filter-type-pill-Update'),
+    ).toBeInTheDocument();
+
     // Click the "×" button on the "App1" filter pill to remove it
     fireEvent.click(screen.getByTestId('active-filter-remove-app-pill-App1'));
-  
+
     // Verify "App1" filter is removed and URL is updated accordingly
     await waitFor(() =>
       expect(navigateMock).toHaveBeenLastCalledWith(
@@ -102,25 +104,31 @@ describe('DenseTable component', () => {
         { replace: true },
       ),
     );
-  
+
     // Verify the "App1" pill is no longer in the document
-    expect(screen.queryByTestId('active-filter-app-pill-App1')).not.toBeInTheDocument();
+    expect(
+      screen.queryByTestId('active-filter-app-pill-App1'),
+    ).not.toBeInTheDocument();
   });
-  
+
   it('clears all filters when "Clear All Filters" button is clicked', async () => {
     render(
       <MemoryRouter initialEntries={['/?filters=type%3AUpdate,app%3AApp1']}>
-        <DenseTable changes={testData} />
+        <ChangeTable changes={testData} />
       </MemoryRouter>,
     );
-  
+
     // Ensure both filter pills are present initially in the filter box
-    expect(screen.getByTestId('active-filter-app-pill-App1')).toBeInTheDocument();
-    expect(screen.getByTestId('active-filter-type-pill-Update')).toBeInTheDocument();
-  
+    expect(
+      screen.getByTestId('active-filter-app-pill-App1'),
+    ).toBeInTheDocument();
+    expect(
+      screen.getByTestId('active-filter-type-pill-Update'),
+    ).toBeInTheDocument();
+
     // Click "Clear All Filters" button
     fireEvent.click(screen.getByTestId('clear-all-filters'));
-  
+
     // Verify all filters are cleared and URL is updated
     await waitFor(() =>
       expect(navigateMock).toHaveBeenLastCalledWith(
@@ -128,34 +136,36 @@ describe('DenseTable component', () => {
         { replace: true },
       ),
     );
-  
+
     // Verify that filter pills are no longer displayed
-    expect(screen.queryByTestId('active-filter-app-pill-App1')).not.toBeInTheDocument();
-    expect(screen.queryByTestId('active-filter-type-pill-Update')).not.toBeInTheDocument();
+    expect(
+      screen.queryByTestId('active-filter-app-pill-App1'),
+    ).not.toBeInTheDocument();
+    expect(
+      screen.queryByTestId('active-filter-type-pill-Update'),
+    ).not.toBeInTheDocument();
   });
-  
+
   it('renders change types and apps as pills with correct styling', () => {
     render(
       <MemoryRouter>
-        <DenseTable changes={testData} />
+        <ChangeTable changes={testData} />
       </MemoryRouter>,
     );
-  
+
     // Find the change type pill by its text content
     const changeTypePill = screen.getByText(/update/i);
     expect(changeTypePill).toBeInTheDocument();
 
-  
     // Find the app pill by its text content
     const appPill = screen.getByText(/app1/i);
     expect(appPill).toBeInTheDocument();
-
   });
 
   it('displays the correct icon for error field', () => {
     render(
       <MemoryRouter>
-        <DenseTable changes={testData} />
+        <ChangeTable changes={testData} />
       </MemoryRouter>,
     );
 
@@ -170,7 +180,7 @@ describe('DenseTable component', () => {
   it('loads initial filters from the URL query string', () => {
     render(
       <MemoryRouter initialEntries={['/?filters=type%3AUpdate,app%3AApp1']}>
-        <DenseTable changes={testData} />
+        <ChangeTable changes={testData} />
       </MemoryRouter>,
     );
 
@@ -182,7 +192,7 @@ describe('DenseTable component', () => {
   it('updates the URL with field-based filters when multiple pills are clicked', async () => {
     render(
       <MemoryRouter>
-        <DenseTable changes={testData} />
+        <ChangeTable changes={testData} />
       </MemoryRouter>,
     );
